Validator compilated

+++ QamarKitoblar.Service/Services/Geners/GenerService.cs
@@ -1,9 +1,11 @@
-﻿using QamarKitoblar.DataAccess.Interfaces.Geners;
+﻿using QamarKitoblar.DataAccess.Interfaces;
+using QamarKitoblar.DataAccess.Interfaces.Geners;
 using QamarKitoblar.DataAccess.Utils;
 using QamarKitoblar.Domain.Entities.Geners;
 using QamarKitoblar.Domain.Exceptions.Geners;
 using QamarKitoblar.Service.Common.Helpers;
 using QamarKitoblar.Service.Dtos.Categories;
+using QamarKitoblar.Service.Interafaces.Common;
 using QamarKitoblar.Service.Interafaces.Geners;
 
 namespace QamarKitoblar.Service.Services.Geners;
@@ -11,10 +13,12 @@ namespace QamarKitoblar.Service.Services.Geners;
 public class GenerService : IGenerService
 {
     private readonly IGenerRepository _generRepository;
+    private readonly IPaginator _paginator;
 
-    public GenerService(IGenerRepository generRepository)
+    public GenerService(IGenerRepository generRepository,IPaginator paginator)
     {
         this._generRepository = generRepository;
+        this._paginator = paginator;
     }
 
     public async Task<long> CountAsync()
@@ -38,6 +42,8 @@ public class GenerService : IGenerService
 
     public async Task<bool> DeleteAsync(long GenerId)
     {
+        var resultDel=await _generRepository.GetByIdAsync(GenerId);
+        if(resultDel is null) { throw new GenerNotFoundException(); }
         var result = await _generRepository.DeleteAsync(GenerId);
         return result > 0;
     }
@@ -45,6 +51,8 @@ public class GenerService : IGenerService
     public async Task<IList<Gener>> GetAllAsync(PaginationParams @params)
     {
         var result = await _generRepository.GetAllAsync(@params);
+        var count = await _generRepository.CountAsync();
+        _paginator.Paginate(count, @params);
         return result;
     }
